requirements.txt
Flask>=2.0.0,<3.0
pandas>=1.3.0,<1.4
matplotlib>=3.4.0,<3.5
networkx>=2.6.0,<2.7
konlpy>=0.5.2,<0.6
JPype1>=1.3.0,<2.0
numpy>=1.21.0,<1.22
seaborn>=0.11.0,<0.12
openpyxl>=3.0.0,<3.1
kiwipiepy>=0.9.0,<0.10
nltk>=3.6.0,<3.7
squarify>=0.4.3,<0.5
wordcloud>=1.8.0,<1.9
Pillow>=8.0.0,<9.0
setuptools>=50.0.0,<71.0
six>=1.15.0,<2.0
gunicorn>=20.0.4,<21.0
